+++ changelog.md
@@ -2,8 +2,15 @@
 
 ## 0.17.0 - 05.11.2020
 
+## new features
+
+- TODO: prefixTypeCategory :: Bool
+
 ### Breaking Changes
 
+- TODO: typeOptions :: f a -> GQLTypeOptions -> GQLTypeOptions
+
+- deexposed constructor of `GQLTypeOptions`.
 - Type name for parametrized types like `One (Two Three)` will be generated directly, concatenating them `OneTwoThree` instead of `One_Two_Three.`
 - Haskell `Float` was renamed to custom scalar type `Float32.`
 - Haskell `Double` now represents GraphQL `Float`.